Splitting out types & using constructor functions
The tokenizer and parser now have independent type files just to keep
the files a bit easier to read and the abstraction levels clearer. The
types may eventually get merged into the implementation files
themselves, but not sure yet.

Constructor functions are now used to remove the need for extra code in
different parsing functions, and to lessen the risk the types come back
in a non-standard way.